Imgui: being able to visualize online
