+++ src/core/progress.h
@@ -6,14 +6,4 @@
 #pragma once
 #include <iostream>
 
-#define PBSTR "||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||"
-#define PBWIDTH 100
-
-void printProgress(int spp_now, int spp_full) {
-    float percentage = float(spp_now + 1) / float(spp_full); 
-    int val = (int) (percentage * 100);
-    int lpad = (int) (percentage * PBWIDTH);
-    int rpad = PBWIDTH - lpad;
-    printf("\r%3d%% (%4d / %4d) [%.*s%*s]", val, spp_now + 1, spp_full, lpad, PBSTR, rpad, "");
-    fflush(stdout);
-}
+void printProgress(int spp_now, int spp_full);
